--- SceneThree.tsx
@@ -1,45 +1,0 @@
-import React, {useRef, useState, Fragment} from 'react';
-import { Scene, WebGLRenderer, PointLight, Mesh } from 'three';
-import { Canvas, useFrame } from 'react-three-fiber';
-
-const Box = (props) => {
-  // This reference will give us direct access to the mesh
-  const mesh = useRef<Mesh>()
-  // Set up state for the hovered and active state
-  const [hovered, setHover] = useState(false)
-  const [active, setActive] = useState(false)
-
-  // Rotate mesh every frame, this is outside of React without overhead
-  useFrame(() => {
-    if (hovered && !active) {
-      mesh.current.rotation.z += 0.01
-      mesh.current.rotation.x += 0.01
-    }
-    if (hovered && active) {
-      mesh.current.rotation.y += 0.02
-      mesh.current.rotation.x += 0.06
-    }
-  })
-
-  return (
-    <mesh
-      {...props}
-      ref={mesh}
-      scale={active ? [1.5, 1.5, 1.5] : [1, 1, 1]}
-      onClick={e => setActive(!active)}
-      onPointerOver={e => setHover(true)}
-      onPointerOut={e => setHover(false)}>
-      <boxBufferGeometry attach="geometry" args={[1, 1, 1]} />
-      <meshStandardMaterial attach="material" color={hovered ? 'hotpink' : 'orange'} />
-    </mesh>
-  )
-}
-
-export default SceneThree => <Fragment>
-    <Canvas>
-      <ambientLight />
-      <pointLight position={[10, 10, 10]} />
-      <Box position={[-2, 0, 0]} />
-      <Box position={[2, 0, 0]} />
-    </Canvas>
-  </Fragment>
