+++ index.tsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { render } from 'react-dom';
 import './style.css';
-import SceneThree from './SceneThree';
+import Area from "./components/3d/Area";
 
 interface AppProps { }
 interface AppState {
@@ -18,7 +18,7 @@ class App extends Component<AppProps, AppState> {
 
   render() {
     return (
-      <SceneThree />
+      <Area />
     );
   }
 }
